move all modules to inventory_utils/
