+++ src/inventory_utils/aliases.py
@@ -1,0 +1,13 @@
+"""
+Purpose: Module of custom type aliases
+Authors: Kenneth Schackart
+"""
+
+import sys
+from typing import List
+
+TaggedBatch = List[List[str]]
+
+# ---------------------------------------------------------------------------
+if __name__ == '__main__':
+    sys.exit('This file is a module, and is not meant to be run.')
