+++ src/ner_final_eval.py
@@ -10,9 +10,11 @@ from typing import BinaryIO, NamedTuple
 
 from torch.utils.data.dataloader import DataLoader
 
-from ner_data_handler import RunParams, get_dataloader
-from utils import (CustomHelpFormatter, save_metrics, get_ner_metrics,
-                   get_ner_model, get_torch_device)
+from inventory_utils.custom_classes import CustomHelpFormatter
+from inventory_utils.filing import get_ner_model, save_metrics
+from inventory_utils.metrics import get_ner_metrics
+from inventory_utils.ner_data_handler import RunParams, get_dataloader
+from inventory_utils.runtime import get_torch_device
 
 
 # ---------------------------------------------------------------------------
